[JENKINS-11643 JENKINS-9258]
Revisiting the defensive check needed for JENKINS-11643 in light of making remember me service works (JENKINS-9258)

I've made changes in the core so that the TokenBasedRememberMeService2.autoLogin consults
the LastGrantedAuthoritiesProperty of the User object in Jenkins 1.556. So when used with
newer version of Jenkins, I can making remember me work with AD.

This fix makes AD plugin behave gracefully with earlier versions, while still allowing me
to leverage new additions in 1.556.