[JENKINS-24195] Ignoring irrelevant Active Directory groups
Added an option to make Jenkins ignore Active Directory
groups that are not being used by the active
Authorization Strategy. This can significantly improve
performance in environments with a large number of groups
but a small number of corresponding rules defined by the
Authorization Strategy.

Performance tests have been made in an enviroment with
Role Based Authorization Strategy, 7000 jobs and
280 AD groups (where only one of the last groups is
used by the AuthorizationStrategy). The average loading
time of an empty List View went from 6.2 seconds
to 0.2 seconds. An sectioned view with 15 sections
of text went from 36 seconds to 0.6 seconds.

When Matrix Based Authorization was used instead,
loading time of the sectioned view went down from
1.2 seconds to 0.1 seconds.