[JENKINS-36248] Provide an ultimate speed option.
If the AD setup only requires Security Groups then we already have these
from the token-groups attribute, so searching memberOf will just make
login take longer.

The exception to this is if you have users are members of a security group
via way of a intermediate distribution group, but for these cases you can
still choose one of the other 2 methods.