the majority of the docker tests did not require static ports
Changed the ports to be dynamically assigned and removed the tests from
the IT profile

Added a assume so that tests would be skipped if docker is not available

Adapted production code to be able to override the DNS server used by
JNDI to aid testing.

CHanged the IT profile into 2 profiles
ITs will run the normal tests *and* the IT tests
onlyITs will only run the ITs (making the pipeline faster as we don't
need to run all tests again inside the customer docker.

When running ITs forkCout will be set to 1 to avoid failures due to port
conflicts.

The WindowsAdsi tests have no test run on CI - they require a very
custom windows environment in order to run.